+++ circuit_maintenance_parser/provider.py
@@ -17,6 +17,7 @@ from circuit_maintenance_parser.parser import EmailDateParser, ICal
 
 from circuit_maintenance_parser.parsers.apple import SubjectParserApple, TextParserApple
 from circuit_maintenance_parser.parsers.aquacomms import HtmlParserAquaComms1, SubjectParserAquaComms1
+from circuit_maintenance_parser.parsers.att import HtmlParserATT1, XlsxParserATT1
 from circuit_maintenance_parser.parsers.aws import SubjectParserAWS1, TextParserAWS1
 from circuit_maintenance_parser.parsers.bso import HtmlParserBSO1
 from circuit_maintenance_parser.parsers.cogent import HtmlParserCogent1, SubjectParserCogent1, TextParserCogent1
@@ -233,6 +234,17 @@ class Arelion(GenericProvider):
     _exclude_filter = PrivateAttr({EMAIL_HEADER_SUBJECT: ["Disturbance Information"]})
 
     _default_organizer = PrivateAttr("support@arelion.com")
+
+
+class ATT(GenericProvider):
+    """ATT provider custom class."""
+
+    _processors: List[GenericProcessor] = PrivateAttr(
+        [
+            CombinedProcessor(data_parsers=[EmailDateParser, HtmlParserATT1, XlsxParserATT1]),
+        ]
+    )
+    _default_organizer = PrivateAttr("g31654@att.com")
 
 
 class AWS(GenericProvider):